+++ backend/core/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     'corsheaders',
     'rest_framework',
 
+    'accounts',
 ]
 
 MIDDLEWARE = [
@@ -122,6 +123,8 @@ CONN_MAX_AGE = 60
 
 
 # ========== AUTH ==========
+AUTH_USER_MODEL = 'accounts.User'
+
 AUTH_PASSWORD_VALIDATORS = [
     {
         'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator',